+++ stochtree/bcf.py
@@ -4406,7 +4406,16 @@ class BCFModel:
 
         raise ValueError(f"term {term} is not a valid BCF model term")
 
-    def summary(self) -> str:
+    def summary(self) -> None:
+        """
+        Summarize a BCF fit with a description of the model that was fit and numeric summaries of any sampled quantities
+
+        Prints summary directly to the console with no return type.
+        
+        Returns
+        -------
+        None
+        """
         # First, print the BCF model
         output_str = "BCF Model Summary:\n"
         output_str += "------------------\n"
@@ -4567,7 +4576,7 @@ class BCFModel:
                 rfx_quantiles = np.quantile(rfx_beta_samples, probs)
                 for p, q in zip(probs, rfx_quantiles):
                     output_str += f"  {p*100:5.1f}%: {q:.3f}\n"
-        return output_str
+        print(output_str)
 
     def __str__(self) -> str:
         if not self.sampled: