+++ CHANGELOG.md
@@ -1,10 +1,16 @@
 # Changelog
 
-# stochtree 0.4.0-dev
+# stochtree 0.4.1
 
 ## New Features
 
 ## Bug Fixes
+
+## Documentation and Other Maintenance
+
+* Converted all R man page titles to title case [#310](https://github.com/StochasticTree/stochtree/pull/310)
+* Updated Python `summary()` methods to print directly to console rather than return a summary string [#310](https://github.com/StochasticTree/stochtree/pull/310)
+* Updated vendored pybind11 dependency to version 3.0.2 [#310](https://github.com/StochasticTree/stochtree/pull/310)
 
 # stochtree 0.4.0